+++ tests/testthat/test-sf.R
@@ -1,43 +1,44 @@
-# context("test-sf.R")
-#
-# data("hsh_sf", package = "gibble")
-# ## a sf version of SpatialPolygons, SpatialLines, SpatialMultiPoints
-# SPS <- hsh_sf[["poly"]]
-# SLS <- hsh_sf[["line"]]
-# MPS <- hsh_sf[["point"]]
-#
-# ## a MULTIPOLYGON
-# P <- SPS[[1]]
-# ## a vector of MULTIPOLYGONS
-# PS <- SPS
-#
-# ## a Line
-# L <- SLS[[1]]
-# ## a Lines
-# LS <- SLS
-#
-# test_that("sf ibble works", {
-#   dSPS <- ibble(SPS)
-#   expect_that(dim(dSPS), equals(c(9L, 5L)))
-#   dSLS <- ibble(SLS, 1)
-#   expect_that(dim(dSLS), equals(c(9L, 4L)))
-#
-#   dMPS <- ibble(MPS)
-#   expect_that(dim(dMPS), equals(c(3L, 4L)))
-#
-#   dP <- ibble(P)
-#   expect_that(dim(dP), equals(c(1L, 4L)))
-#
-#   dPS <- ibble(PS)
-#   expect_that(dim(dPS), equals(c(6L, 5L)))
-#
-#   dL <- ibble(L)
-#   expect_that(dim(dL), equals(c(1L, 3L)))
-#
-#   dLS <- ibble(LS)
-#   expect_that(dim(dLS), equals(c(6L, 5L)))
-#
-# })
+context("test-sf.R")
+
+data("hsh_sf", package = "gibble")
+## a sf version of SpatialPolygons, SpatialLines, SpatialMultiPoints
+SPS <- hsh_sf[["poly"]]
+SLS <- hsh_sf[["line"]]
+MPS <- hsh_sf[["point"]]
+
+## a MULTIPOLYGON
+P <- SPS[[1]]
+## a vector of MULTIPOLYGONS
+PS <- SPS
+
+## a Line
+L <- SLS[[1]]
+## a Lines
+LS <- SLS
+
+test_that("sf ibble works", {
+  dSPS <- ibble(SPS)
+  expect_that(dim(dSPS), equals(c(9L, 5L)))
+  dSLS <- ibble(SLS, 1)
+  expect_that(dim(dSLS), equals(c(9L, 5L)))
+
+  dMPS <- ibble(MPS)
+  expect_that(dim(dMPS), equals(c(50L, 4L)))
+
+  # dP <- ibble(P)
+  # expect_that(dim(dP), equals(c(1L, 4L)))
+  #
+  # dPS <- ibble(PS)
+  # expect_that(dim(dPS), equals(c(6L, 5L)))
+  #
+  # dL <- ibble(L)
+  # expect_that(dim(dL), equals(c(1L, 3L)))
+  #
+  # dLS <- ibble(LS)
+  # expect_that(dim(dLS), equals(c(6L, 5L)))
+
+})
+
 #
 #
 # test_that("sp gibble methods exist", {